backend/requirements.txt
websockets
psycopg2-binary
fastapi
uvicorn[standard]
pandas
statsmodels
numpy